Added data_prep, updated data_retrieval and splitting_data
Data prep cleans and merges dat for modelling. Cleaning and merging of passengers data still needs to be fixed. Data retrievl was updated to include retrieval of flight_test table, and to not interfere with the flight_100k_random csv that is used for analysis in the exploratory_analysis notebook. Splitting data had the splitting of fuel data removed, and was updated to take in the cleaned feature table from data prep, while still maintaining file output names as to not disrupt work in modelling notebook.